+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## 1.1.0
 
-* Add [Just](https://github.com/casey/just) and cargo tarpaulin for tests coverage
+* Add [Just](https://github.com/casey/just) and [cargo tarpaulin](https://crates.io/crates/cargo-tarpaulin)
+* Add minimal tests with [Testinfra](https://testinfra.readthedocs.io/en/latest/)
+* Remove not needed cargo workspaces and wasm-tools
 
 ## 1.0.0